18/20

+++ mark.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 
-note=20
-DOSSIER=$1
-
-IFS='_' read -r -a tableau <<< "$DOSSIER"
-nom=${tableau[0]}
-prenom=${tableau[1]}
-
+note=$(./note.sh)
 malus=$(./malus.sh)
 note=$((note + malus))
+
+if [ ! -f readme.txt ]; then
+    echo "Fichier readme.txt manquant (format: Prenom Nom)"
+    exit 1
+fi
+
+read prenom nom < readme.txt
 
 echo "Note finale de $prenom $nom : $note/20"
 
 if [ ! -f "note.csv" ]; then
     echo "Nom,Prénom,Note" > "note.csv"
 fi
-echo "'$nom','$prenom',$note" >> "note.csv"
+
+echo "$nom,$prenom,$note" >> "note.csv"
